+++ server/package.json
@@ -2,20 +2,27 @@
   "name": "server",
   "version": "1.0.0",
   "description": "",
-  "main": "app.ts",
+  "main": "src/app.ts",
   "scripts": {
-    "start": "ts-node app.ts"
+    "start": "nodemon"
   },
   "author": "",
   "license": "ISC",
   "dependencies": {
     "cors": "^2.8.5",
+    "crypto": "^1.0.1",
     "express": "^4.18.2",
+    "jsrsasign": "^11.1.0",
+    "morgan": "^1.10.0",
+    "node-forge": "^1.3.1",
     "ts-node": "^10.9.2",
     "typescript": "^5.3.3"
   },
   "devDependencies": {
     "@types/cors": "^2.8.17",
-    "@types/express": "^4.17.21"
+    "@types/express": "^4.17.21",
+    "@types/jsrsasign": "^10.5.14",
+    "@types/morgan": "^1.9.9",
+    "nodemon": "^3.1.7"
   }
 }
